move header componetns to own feature
