+++ src/app/domains/individual/feature-edit/individual-edit.page.ts
@@ -5,7 +5,6 @@ import { TitleService } from '@core/services/title.service';
 import { TranslateModule } from '@ngx-translate/core';
 import { Individual } from '@shared/models/individual.model';
 import { IndividualService } from '../../../shared/services/individual.service';
-import { IndividualHeaderComponent } from '../shared/individual-header.component';
 import { IndividualEditViewComponent } from './individual-edit-form.widget';
 import { IndividualEditHeaderComponent } from './individual-edit-header.widget';
 
@@ -13,7 +12,7 @@ import { IndividualEditHeaderComponent } from './individual-edit-header.widget';
   templateUrl: './individual-edit.page.html',
   styleUrls: ['./individual-edit.page.scss'],
   standalone: true,
-  imports: [IndividualHeaderComponent, TranslateModule, IndividualEditViewComponent, IndividualEditHeaderComponent]
+  imports: [TranslateModule, IndividualEditViewComponent, IndividualEditHeaderComponent]
 })
 export class IndividualEditComponent extends BaseDetailComponent<Individual> implements OnInit {
   isNewIndividual: boolean;
