+++ src/app/domains/individual/feature-header/individual-header-actions.component.scss
