+++ src/app/domains/individual/feature-header/individual-header-graph.widget.scss
