+++ src/app/domains/individual/feature-header/individual-header-map.widget.scss
