+++ src/app/domains/individual/feature-header/individual-header.model.ts
