+++ src/app/domains/individual/individual.page.ts
@@ -13,9 +13,9 @@ import { IndividualService } from '../../shared/services/individual.service';
 import { IndividualDescriptionButtonsComponent } from './feature-buttons/individual-buttons.widget';
 import { IndividualDescriptionHeaderComponent } from './feature-description/individual-info.widget';
 import { IndividualDescriptionBasicInfoComponent } from './feature-description/masterdata-info.component';
+import { IndividualHeaderComponent } from './feature-header/individual-header.component';
 import { ObservationViewComponent } from './feature-observations/individual-observation.widget';
 import { IndividualSubscriptionButtonComponent } from './feature-subscription/individual-subscription.widget';
-import { IndividualHeaderComponent } from './shared/individual-header.component';
 
 @Component({
   templateUrl: './individual.page.html',
